sync from upstream

+++ es-app/src/guis/GuiFastSelect.cpp
@@ -1,5 +1,6 @@
 #include "guis/GuiFastSelect.h"
-#include "ThemeData.h"
+
+#include "views/gamelist/IGameListView.h"
 #include "FileSorts.h"
 #include "SystemData.h"
 
@@ -79,7 +80,7 @@ bool GuiFastSelect::input(InputConfig* config, Input input)
 	{
 		mSortId--;
 		if(mSortId < 0)
-			mSortId += FileSorts::SortTypes.size();
+			mSortId += (int)FileSorts::SortTypes.size();
 
 		updateSortText();
 		return true;
@@ -115,7 +116,7 @@ void GuiFastSelect::scroll()
 	mLetterId += mScrollDir;
 	if(mLetterId < 0)
 		mLetterId += LETTERS.length();
-	else if(mLetterId >= (int)LETTERS.length())
+	else if(mLetterId >= LETTERS.length())
 		mLetterId -= LETTERS.length();
 
 	mLetterText.setText(LETTERS.substr(mLetterId, 1));
